Added unit tests for DroneServiceImpl

+++ src/main/java/com/drones/domain/Medication.java
@@ -24,7 +24,7 @@ public class Medication {
     private Long id;
     @Column(unique = true)
     private String name;
-    private Double weight;
+    private Integer weight;
     private String code;
     private String image;
 }
